miraikakakubatch/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miraikakaku-batch"
version = "1.0.0"
description = "MiraiKakaku Batch Processing"
authors = [
    {name = "MiraiKakaku Team", email = "dev@miraikakaku.com"}
]
dependencies = [
    "functions-framework>=3.0.0",
    "psycopg2-binary>=2.9.9",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "yfinance>=0.2.0"
]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["functions"]

[tool.pylint.messages_control]
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "import-error"
]

[tool.pylint.format]
max-line-length = 88